apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
fpdf2>=2.7.0
